requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
faiss-cpu
requests
langchain-openai
streamlit
langchain_ollama